Make the ScrapingClient somewhat api-compatible
It's not going to be perfect, but the idea is that for the most basic of
cases it should be a pretty close replacement. The goal is to keep the
amount of work to support both API and scraping-based clients to a minimum.

To support this, the WebClient now uses delegation instead of
inheritance to add scraper-based functionality. This enables the
`ScrapingClient` class to use the same function names without
automatically overriding the `stravalib.Client` functions when used
through the `WebClient` class.